Fix loss of precision

+++ osu.Framework.Tests/Visual/Input/TestSceneInputResampler.cs
@@ -165,9 +165,9 @@ namespace osu.Framework.Tests.Visual.Input
                 Texture = texture;
                 Colour = colour;
 
-                const int target_raw = 1024;
+                const float target_raw = 1024;
 
-                for (int i = 0; i < target_raw; i++)
+                for (float i = 0; i < target_raw; i++)
                 {
                     float x = (MathF.Sin(i / target_raw * (MathF.PI * 0.5f)) * 200) + 50.5f;
                     float y = (MathF.Cos(i / target_raw * (MathF.PI * 0.5f)) * 200) + 50.5f;
